pyproject.toml
[tool.poetry]
name = "advisoryopinions"
version = "0.1.0"
description = ""
authors = ["Tim Showers <showerst@gmail.com>"]
license = "GPL-v3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
lxml = "^4.9.3"
python-dateutil = "^2.8.2"
dataclasses-json = "^0.6.2"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
scrape = "advisoryopinions.scrape:scrape"